Refactor VPN Gateway and unit test, add VPC -HAS-> VPN Gateway relationship
