+++ docs/spec/src/index.ts
@@ -3,7 +3,6 @@ import { IntegrationSpecConfig } from '@jupiterone/integration-sdk-core';
 import { IntegrationConfig } from '../../../src/config';
 import { albSpec } from './alb';
 import { ecsSpec } from './ecs';
-import { vpnSpec } from './vpn';
 import { vpcSpec } from './vpc';
 import { autoScalingSpec } from './auto-scaling';
 import { ramSpec } from './ram';
@@ -13,7 +12,6 @@ export const invocationConfig: IntegrationSpecConfig<IntegrationConfig> = {
     ...ecsSpec,
     ...albSpec,
     ...ramSpec,
-    ...vpnSpec,
     ...vpcSpec,
     ...autoScalingSpec,
   ],
