+++ docs/spec/src/vpc/index.ts
@@ -40,6 +40,24 @@ export const vpcSpec: StepSpec<IntegrationConfig>[] = [
   },
   {
     /**
+     * Action: DescribeVpnGateways
+     * PATTERN: Fetch Entities
+     */
+    id: 'fetch-vpn-gateways',
+    name: 'Fetch VPN Gateways',
+    entities: [
+      {
+        resourceName: 'VPN Gateway',
+        _type: 'alibaba_cloud_vpn_gateway',
+        _class: ['Gateway'],
+      },
+    ],
+    relationships: [],
+    dependsOn: [],
+    implemented: true,
+  },
+  {
+    /**
      * PATTERN: Build Child Relationships
      */
     id: 'build-vpc-has-nat-gateway-relationships',
@@ -54,6 +72,24 @@ export const vpcSpec: StepSpec<IntegrationConfig>[] = [
       },
     ],
     dependsOn: ['fetch-vpcs', 'fetch-nat-gateways'],
+    implemented: true,
+  },
+  {
+    /**
+     * PATTERN: Build Child Relationships
+     */
+    id: 'build-vpc-has-vpn-gateway-relationships',
+    name: 'Build VPC has VPN Gateway Relationships',
+    entities: [],
+    relationships: [
+      {
+        _type: 'alibaba_cloud_vpc_has_vpn_gateway',
+        sourceType: 'alibaba_cloud_vpc',
+        _class: RelationshipClass.HAS,
+        targetType: 'alibaba_cloud_vpn_gateway',
+      },
+    ],
+    dependsOn: ['fetch-vpcs', 'fetch-vpn-gateways'],
     implemented: true,
   },
 ];
