+++ src/steps/index.ts
@@ -1,6 +1,5 @@
 import { albSteps } from './alb/index';
 import { ecsSteps } from './ecs/index';
-import { vpnSteps } from './vpn';
 import { vpcSteps } from './vpc';
 import { autoScalingSteps } from './auto-scaling';
 import { ramSteps } from './ram/index';
@@ -9,7 +8,6 @@ const integrationSteps = [
   ...ecsSteps,
   ...albSteps,
   ...ramSteps,
-  ...vpnSteps,
   ...vpcSteps,
   ...autoScalingSteps,
 ];
