+++ src/steps/vpc/index.ts
@@ -12,9 +12,10 @@ import { VPCEntities, VPCRelationships, VPCSteps } from './constants';
 import {
   createNATGatewayEntity,
   createVPCEntity,
+  createVPNGatewayEntity,
   getVpcKey,
 } from './converter';
-import { NATGateway } from './types';
+import { NATGateway, VPNGateway } from './types';
 
 export async function fetchVPCs({
   logger,
@@ -37,6 +38,18 @@ export async function fetchNATGateways({
 
   await client.iterateNATGateways(async (natGateway) => {
     await jobState.addEntity(createNATGatewayEntity(natGateway));
+  });
+}
+
+export async function fetchVPNGateways({
+  logger,
+  instance,
+  jobState,
+}: IntegrationStepExecutionContext<IntegrationConfig>): Promise<void> {
+  const client = createVPCClient(instance.config, logger);
+
+  await client.iterateVPNGateways(async (vpnGateway) => {
+    await jobState.addEntity(createVPNGatewayEntity(vpnGateway));
   });
 }
 
@@ -75,6 +88,41 @@ export async function buildVPCHasNATGatewayRelationship({
   );
 }
 
+export async function buildVPCHasVPNGatewayRelationship({
+  logger,
+  jobState,
+}: IntegrationStepExecutionContext<IntegrationConfig>) {
+  await jobState.iterateEntities(
+    { _type: VPCEntities.VPN_GATEWAY._type },
+    async (vpnGatewayEntity) => {
+      const vpnGateway = getRawData<VPNGateway>(vpnGatewayEntity);
+      if (!vpnGateway) {
+        logger.warn(
+          { _key: vpnGatewayEntity._key },
+          'Could not get raw data for VPN Gateway entity',
+        );
+        return;
+      }
+
+      const vpcId = vpnGateway.VpcId;
+      if (!vpcId) {
+        return;
+      }
+
+      const vpcEntity = await jobState.findEntity(getVpcKey(vpcId));
+      if (vpcEntity) {
+        await jobState.addRelationship(
+          createDirectRelationship({
+            _class: RelationshipClass.HAS,
+            from: vpcEntity,
+            to: vpnGatewayEntity,
+          }),
+        );
+      }
+    },
+  );
+}
+
 export const vpcSteps: IntegrationStep<IntegrationConfig>[] = [
   {
     id: VPCSteps.FETCH_VPCS,
@@ -93,11 +141,27 @@ export const vpcSteps: IntegrationStep<IntegrationConfig>[] = [
     executionHandler: fetchNATGateways,
   },
   {
+    id: VPCSteps.FETCH_VPN_GATEWAYS,
+    name: 'Fetch VPN Gateways',
+    entities: [VPCEntities.VPN_GATEWAY],
+    relationships: [],
+    dependsOn: [],
+    executionHandler: fetchVPNGateways,
+  },
+  {
     id: VPCSteps.BUILD_VPC_NAT_GATEWAY_RELATIONSHIPS,
     name: 'Build VPC has NAT Gateway Relationships',
     entities: [],
     relationships: [VPCRelationships.VPC_HAS_NAT_GATEWAY],
     dependsOn: [VPCSteps.FETCH_VPCS, VPCSteps.FETCH_NAT_GATEWAYS],
     executionHandler: buildVPCHasNATGatewayRelationship,
+  },
+  {
+    id: VPCSteps.BUILD_VPC_VPN_GATEWAY_RELATIONSHIPS,
+    name: 'Build VPC has VPN Gateway Relationships',
+    entities: [],
+    relationships: [VPCRelationships.VPC_HAS_VPN_GATEWAY],
+    dependsOn: [VPCSteps.FETCH_VPCS, VPCSteps.FETCH_VPN_GATEWAYS],
+    executionHandler: buildVPCHasVPNGatewayRelationship,
   },
 ];
