+++ src/steps/vpc/types.ts
@@ -134,3 +134,21 @@ export interface NATGateway {
   PrivateLinkEnabled?: boolean;
   PrivateLinkMode?: string;
 }
+
+export interface VPNGateway {
+  Status?: string;
+  IpsecVpn?: string;
+  SslVpn?: string;
+  VpnType?: string;
+  EnableBgp?: boolean;
+  CreateTime?: number;
+  EndTime?: number;
+  BusinessStatus?: string;
+  Name?: string;
+  VpcId?: string;
+  VpnGatewayId: string;
+  ChargeType?: string;
+  NetworkType?: string;
+  Spec?: string;
+  Tag?: string;
+}
