--- src/steps/vpn/client/types/response.ts
@@ -1,8 +1,0 @@
-import { VPNGateway } from '../../types';
-import { PaginatedResponse } from '../../../../client/types/response';
-
-export interface DescribeVpnGatewaysResponse extends PaginatedResponse {
-  VpnGateways: {
-    VpnGateway: VPNGateway[];
-  };
-}
